auth: fix improper logout when access token expired If the user refreshed the brower while the access token was expired but the session token wasn't the app would improperly log the user out.
This patch fixes the behavior so the access token is refreshed if the session token is still valid.

Introduced in: https://github.com/gregtwallace/certwarden-frontend/commit/19d2bdca6d8ede7731d014c393213aae590356b2

Caused by parse requiring both the access and session tokens to be valid when getAuth() was called. This requirement should only be imposed when fetching a new auth from the backend.